style(clippy:pedantic): pedantic style fixes

+++ src/dotfiles.rs
@@ -1,36 +1,35 @@
+use std::fmt;
 use std::fs;
 use std::path::Path;
-use std::fmt;
 
 use dirs::home_dir;
 
 #[cfg(target_os = "linux")]
-use std::os::unix::fs::symlink as symlink;
+use std::os::unix::fs::symlink;
 
 #[cfg(target_os = "windows")]
-use std::os::windows::fs::symlink as symlink;
+use std::os::windows::fs::symlink;
 #[derive(Debug)]
 enum Op {
-    SYMFILE,
-    SYMDIR,
-    INVALID,
+    Symfile,
+    Symdir,
+    Invalid,
 }
 
 #[derive(Debug)]
 enum FormatFile {
     OrgTable,
-    CSV,
+    Csv,
 }
 
 impl FormatFile {
     fn parse(format_str: &str) -> Result<Self, String> {
         match format_str {
             "org" => Ok(FormatFile::OrgTable),
-            "csv" => Ok(FormatFile::CSV),
-            _ => Err(format!("Invalid format {}.", format_str)),
+            "csv" => Ok(FormatFile::Csv),
+            _ => Err(format!("Invalid format {format_str}.")),
         }
     }
-
 }
 #[derive(Debug)]
 struct Dot {
@@ -38,19 +37,19 @@ struct Dot {
     name: String,
     source: String,
     dest: String,
-    operation: Op
+    operation: Op,
 }
 
 #[allow(dead_code)]
 pub struct Flags {
+    file_format: FormatFile,
     headers: bool,
     force: bool,
-    file_format: FormatFile,    
 }
 
 impl Flags {
-    pub fn build(file_format: String, headers: bool, force: bool) -> Self {
-        let file_format = FormatFile::parse(file_format.as_str()).unwrap();
+    pub fn build(file_format: &str, headers: bool, force: bool) -> Self {
+        let file_format = FormatFile::parse(file_format).unwrap();
         Self {
             file_format,
             headers,
@@ -62,47 +61,50 @@ impl Flags {
 pub struct Dots {
     filename: String,
     flags: Flags,
-    dots: Vec<Dot>,
+    dotfiles: Vec<Dot>,
 }
 
 impl fmt::Display for Dot {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        write!(f, "[{}:{}] {} -> {} | {:?}",
-            self.name,
-            self.line,
-            self.source,
-            self.dest,
-            self.operation,
+        write!(
+            f,
+            "[{}:{}] {} -> {} | {:?}",
+            self.name, self.line, self.source, self.dest, self.operation,
         )
     }
 }
 
 impl fmt::Display for Dots {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        let mut display = String::from("");
-        self.dots.iter().for_each(|dot| display.push_str(&*format!("{}\n", dot)));
-        write!(f, "{}",  display)
+        let mut display = String::new();
+        self.dotfiles
+            .iter()
+            .for_each(|dot| display.push_str(&format!("{dot}\n")));
+        write!(f, "{display}")
     }
 }
 
 impl Dots {
-    pub fn build(flags: Flags) -> Result<Self, String> {
-        Ok(Self {
-            filename: "".to_string(),
+    pub fn build(flags: Flags) -> Self {
+        Self {
+            filename: String::new(),
             flags,
-            dots: vec![],
-        })
+            dotfiles: vec![],
+        }
     }
 
     pub fn parse_file(&mut self, filename: &str) -> Result<(), String> {
         let filepath = Path::new(filename).canonicalize().unwrap();
-        let file_contents = fs::read_to_string(filepath).expect(&*format!("File {} does not exist", filename));
+        let file_contents = fs::read_to_string(filepath)
+            .unwrap_or_else(|_| panic!("File {filename} does not exist"));
         match self.flags.file_format {
             FormatFile::OrgTable => {
                 let skip_headers = if self.flags.headers { 2 } else { 0 };
-                for (num,line) in file_contents.lines().skip(skip_headers).enumerate() {
-                    let values = line.split("|").collect::<Vec<_>>();
-                    if values.len() < 5 {continue};
+                for (num, line) in file_contents.lines().skip(skip_headers).enumerate() {
+                    let values = line.split('|').collect::<Vec<_>>();
+                    if values.len() < 5 {
+                        continue;
+                    };
 
                     let dot = Dot::new(
                         num,       // Line number
@@ -111,59 +113,56 @@ impl Dots {
                         values[3], // Dest
                         values[4], // Operation
                     );
-                    self.dots.push(dot);
+                    self.dotfiles.push(dot);
                 }
                 self.filename = filename.to_string();
                 Ok(())
-            },
-            FormatFile::CSV => todo!("ERROR: FormatFile::CSV is not implemented yet."),
+            }
+            FormatFile::Csv => Err(String::from("ERROR: FormatFile::Csv is not implemented yet.")),
             // _ => Err(format!("Invalid file type dotfiles declaration: {:?}", self.file_format)),
         }
     }
 
     fn format_error(filename: &String, line: usize, name: &String, message: &String) -> String {
-        format!("./{}:{}:0 -> Dotfile `{}`\n\t{}",
-            filename,
-            line,
-            name,
-            message)
+        format!(
+            "./{filename}:{line}:0 -> Dotfile `{name}`\n\t{message}"
+        )
     }
 
     pub fn verify_dotfiles(&self) -> Result<(), String> {
-        let errors: Vec<Result<(), String>> = self.dots.iter()
+        let errors: Vec<Result<(), String>> = self.dotfiles.iter()
             .map(|dot|{
-                let mut err_str = String::from("");
+                let mut err_str = String::new();
                 let mut has_errors = false;
                 let source_path = Path::new(&*(dot.source));
                 let dest_path = Path::new(&*(dot.dest));
-                let line = dot.line + {if self.flags.headers {2 + 1} else {0 + 1}} ;
+                let line = dot.line + {if self.flags.headers {2 + 1} else {1}} ;
                 // Source path
                 match dot.operation {
-                    Op::SYMFILE => {
+                    Op::Symfile => {
                         if !source_path.is_file() {
                             has_errors = true;
                             let error_msg = format!("Path to the source file {} is not valid", dot.source);
                             err_str.push_str(
-                                Dots::format_error(&self.filename, line, &dot.name, &error_msg).as_str())
+                                Dots::format_error(&self.filename, line, &dot.name, &error_msg).as_str());
                         }
                     },
-                    Op::SYMDIR => {
+                    Op::Symdir => {
                         if !source_path.is_dir() {
                             has_errors = true;
                             let error_msg = format!("Path to the source directory {} is not valid", dot.source);
                             err_str.push_str(
                                 Dots::format_error(&self.filename, line, &dot.name, &error_msg).as_str()
-                            )
+                            );
                         }
                     },
-                    Op::INVALID => {
+                    Op::Invalid => {
                         has_errors = true;
-                        let error_msg = format!("Invalid operation. Only allowed `SYMFILE` or `SYMDIR` for files or directories respectively");
+                        let error_msg = "Invalid operation. Only allowed `Symfile` or `Symdir` for files or directories respectively".to_string();
                         err_str.push_str(
                             Dots::format_error(&self.filename, line, &dot.name, &error_msg).as_str()
-                        )
+                        );
                     }
-                    
                 }
 
                 // Destination path
@@ -172,61 +171,62 @@ impl Dots {
                     let error_msg = format!("Path to destination {} is not a valid", dot.dest);
                     err_str.push_str(
                         Dots::format_error(&self.filename, line, &dot.name, &error_msg).as_str()
-                    )
+                    );
                 }
-                
-                if !has_errors { Ok(()) } else {Err(err_str)}
+                if has_errors { Err(err_str) } else {Ok(())}
 
             }).collect::<Vec<_>>();
-        
+
         let mut has_errors = false;
         for d in errors {
             if let Err(err) = d {
                 has_errors = true;
-                eprintln!("ERROR: {}", err)
+                eprintln!("ERROR: {err}");
             }
-        };
-        
-        if has_errors { Err("Dotfiles contains errors.".to_string())} else {Ok(())}
+        }
+        if has_errors {
+            Err("Dotfiles contains errors.".to_string())
+        } else {
+            Ok(())
+        }
     }
 
     pub fn execute(&self) {
-        self.dots.iter().for_each(|dot| {
+        self.dotfiles.iter().for_each(|dot| {
             let _ = dot.execute(&self.flags);
-        })
+        });
     }
 }
 
 impl Dot {
-    fn new(line: usize,  name: &str, source: &str, dest: &str, op: &str) -> Self {
+    fn new(line: usize, name: &str, source: &str, dest: &str, op: &str) -> Self {
         let name = name.trim().to_string();
         let source = source.trim().to_string();
         let dest = dest.trim().to_string();
         let op = op.trim().to_string();
 
-        let dest = match Path::new(dest.as_str()).is_absolute() {
-            true => {
-                Dot::strip_dir(dest)
-            },
-            false => {
-                let dest = Dot::strip_dir(dest);
-                format!("{}/{}", home_dir().unwrap().display(), dest)
-            }
+        let dest = if Path::new(dest.as_str()).is_absolute() {
+            Dot::strip_dir(dest)
+        }
+        else {
+            let dest = Dot::strip_dir(dest);
+            format!("{}/{}", home_dir().unwrap().display(), dest)
         };
         
+
         let operation = match op.to_lowercase().as_str() {
-            "symfile" => Op::SYMFILE,
-            "symdir"  => Op::SYMDIR,
-            _ => Op::INVALID,
+            "symfile" => Op::Symfile,
+            "symdir" => Op::Symdir,
+            _ => Op::Invalid,
         };
-        
+
         Self {
             line,
             name,
             source,
             dest,
             operation,
-        } 
+        }
     }
 
     fn execute(&self, flags: &Flags) -> Result<(), String> {
@@ -238,42 +238,46 @@ impl Dot {
         let _ = fs::create_dir_all(dest_path.parent().unwrap());
         if flags.force && dest_path.exists() {
             if let Ok(meta) = dest_path.metadata() {
-                assert!(dest_path.exists(), "IMPOSSIBLE: Dest path exists here");                if meta.is_dir() {
+                assert!(dest_path.exists(), "IMPOSSIBLE: Dest path exists here");
+                if meta.is_dir() {
                     fs::remove_dir_all(dest_path).unwrap();
                 } else if meta.is_file() || meta.is_symlink() {
                     fs::remove_file(dest_path).unwrap();
-                }else {
+                } else {
                     panic!("WTF is this file {}", dest_path.display());
                 }
             }
-            
         }
         match symlink(source_path, dest_path) {
-            Ok(_) => {
-                println!("[LOG] Source {} to dest {}",
-                    &self.source,
-                    &self.dest,
-                );
+            Ok(()) => {
+                println!("[LOG] Source {} to dest {}", &self.source, &self.dest,);
                 Ok(())
             }
-,
             Err(err) => {
-                let err_msg = format!("Error {} while symlinking {} to {} ({:?})",
-                    err, &self.source, &self.dest,
-                    self.operation);
-                eprintln!("{}", Dots::format_error(&"".to_string(), self.line, &self.name, &err_msg));
+                let err_msg = format!(
+                    "Error {} while symlinking {} to {} ({:?})",
+                    err, &self.source, &self.dest, self.operation
+                );
+                eprintln!(
+                    "{}",
+                    Dots::format_error(&String::new(), self.line, &self.name, &err_msg)
+                );
+                
                 Err("Error while symlinking".to_string())
-            },
+            }
         }
-
     }
 
     fn strip_dir(dir: String) -> String {
-        let dir = if let Some(d) = &dir.strip_suffix("/"){
-            if dir.len() > 1 {d.to_string()} else {dir}
-        } else {dir};
+        let dir = if let Some(d) = &dir.strip_suffix("/") {
+            if dir.len() > 1 {
+                (*d).to_string()
+            } else {
+                dir
+            }
+        } else {
+            dir
+        };
         dir
     }
 }
-
-
